requirements.txt
python-telegram-bot==20.8
PyMuPDF==1.23.8
markdownify==0.11.6
langchain==0.1.9
openai==1.12.0
chromadb==0.4.22
python-dotenv==1.0.0
tiktoken==0.5.2
python-docx==0.8.11
docx2txt==0.8
tqdm==4.66.1 